// web/src/generated/lvrt_oracle/types/signerKind.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export const SignerKind = {
  0: "Pusher",
  1: "Enclave",
  2: "Receipts",
  Pusher: 0,
  Enclave: 1,
  Receipts: 2,
} as const;

export type SignerKind = (typeof SignerKind)[Exclude<
  keyof typeof SignerKind,
  number
>];

export type SignerKindArgs = SignerKind;

export function getSignerKindEncoder(): FixedSizeEncoder<SignerKindArgs> {
  return getEnumEncoder(SignerKind);
}

export function getSignerKindDecoder(): FixedSizeDecoder<SignerKind> {
  return getEnumDecoder(SignerKind as Omit<typeof SignerKind, number>);
}

export function getSignerKindCodec(): FixedSizeCodec<
  SignerKindArgs,
  SignerKind
> {
  return combineCodec(getSignerKindEncoder(), getSignerKindDecoder());
}