// web/src/generated/lvrt_power/types/powerKind.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export const PowerKind = {
  0: "Squared",
  1: "Ratio",
  Squared: 0,
  Ratio: 1,
} as const;

export type PowerKind = (typeof PowerKind)[Exclude<
  keyof typeof PowerKind,
  number
>];

export type PowerKindArgs = PowerKind;

export function getPowerKindEncoder(): FixedSizeEncoder<PowerKindArgs> {
  return getEnumEncoder(PowerKind);
}

export function getPowerKindDecoder(): FixedSizeDecoder<PowerKind> {
  return getEnumDecoder(PowerKind as Omit<typeof PowerKind, number>);
}

export function getPowerKindCodec(): FixedSizeCodec<PowerKindArgs, PowerKind> {
  return combineCodec(getPowerKindEncoder(), getPowerKindDecoder());
}